+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Local representation of performance assets (#778)
 - Use localrep data in tasks views (#760)
 - Use localrep performances in tasks views (#780)
+- Localrep computeplan status field, that is synced when recieving computetask event update instead of in the computeplan view  (#762)
 
 ### Changed
 - When resyncing the local representation, only fetch the latest events (#656)